Change priors for available parameters

Change priors for available parameters

set_prior(x, ...)

Arguments

...
other arguments passed to function. In the case of intercept and coefficients, these are 'mu' and 'sd' (prior means and variances). In the case of the residual variance, the rate and shape/scale parameter must be passed as 'alpha' and 'beta'

par
name of the coefficient, given as b0 (intercept) or b1, ... for coefficients.

Value

updated blm object with new priors

Examples

bfit <- set_prior(bfit,"b0", mu=40, sd=7)

#> Error in set_prior(bfit, "b0", mu = 40, sd = 7): object 'bfit' not found

bfit <- set_prior(bfit, "sigma", alpha=2, beta=1)

#> Error in set_prior(bfit, "sigma", alpha = 2, beta = 1): object 'bfit' not found